agregando botones de alerta para carrito de compras

+++ src/app/components/cart-app/cart-app.component.ts
@@ -4,6 +4,8 @@ import { CartItem } from '../../models/cartitem';
 import { NavbarComponent } from '../navbar/navbar.component';
 import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../../services/sharing-data.service';
+import Swal from 'sweetalert2'
+
 
 @Component({
   selector: 'cart-app',
@@ -50,24 +52,47 @@ export class CartAppComponent implements OnInit {
       this.router.navigate(['/cart'], {
         state: { items: this.items, total: this.total }
       });
+      Swal.fire({
+        title: "Shopping Cart",
+        text: "Nuevo producto agregado al carro",
+        icon: "success"
+      });
     });
   }
 
   onDeleteCart(): void {
     this.sharingDataService.idProductEventEmitter.subscribe(id => {
-      this.items = this.items.filter(item => item.product.id !== id);
-      if (this.items.length == 0) {
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
-      
-      this.router.navigateByUrl('/', { skipLocationChange: true}).then(()=>{
-        this.router.navigate(['/cart'], {
-          state: { items: this.items, total: this.total }
-        });
-      })
+      Swal.fire({
+        title: "Estas seguro que deseas eliminar el item?",
+        text: "El item se eliminará del carro de compras!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Si, estoy seguro!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter(item => item.product.id !== id);
+          if (this.items.length == 0) {
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
+          
+          this.router.navigateByUrl('/', { skipLocationChange: true}).then(()=>{
+            this.router.navigate(['/cart'], {
+              state: { items: this.items, total: this.total }
+            });
+          })
+          Swal.fire({
+            title: "Eliminado!",
+            text: "Se ha eliminado el item del carrito de compras.",
+            icon: "success"
+          });
+        }
+      });
+
     })
   }
